+++ conjunction/lib/conjunction.rb
@@ -2,7 +2,6 @@
 
 require "active_support"
 require "directive"
-require "spicerack"
 
 require "conjunction/version"