+++ directive/lib/directive/config_object.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
-require "spicerack"
+require "active_model"
+require "substance/input_object"
+require "tablesalt/stringable_object"
 
 module Directive
-  class ConfigObject < Spicerack::InputObject
+  class ConfigObject < Substance::InputObject
     include Singleton
     include ActiveModel::AttributeAssignment