+++ directive/lib/directive/spec_helper/matchers/configuration/define_config_option.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "spicerack/rspec/custom_matchers/define_option"
+require "substance/rspec/custom_matchers/define_option"
 
 module Directive
   module SpecHelper